логика обновления перенесена в smart
